+++ tools/c7n_azure/c7n_azure/resources/arm.py
@@ -13,7 +13,8 @@
 # limitations under the License.
 
 import six
-from c7n_azure.actions import Tag, AutoTagUser, RemoveTag, TagTrim, TagDelayedAction, DeleteAction
+from c7n_azure.actions.tagging import Tag, AutoTagUser, RemoveTag, TagTrim, TagDelayedAction
+from c7n_azure.actions.delete import DeleteAction
 from c7n_azure.filters import (MetricFilter, TagActionFilter,
                                DiagnosticSettingsFilter, PolicyCompliantFilter)
 from c7n_azure.provider import resources
