+++ tools/c7n_azure/c7n_azure/resources/key_vault.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from azure.graphrbac import GraphRbacManagementClient
-from c7n_azure.actions import AzureBaseAction
+from c7n_azure.actions.base import AzureBaseAction
 from c7n_azure.provider import resources
 from c7n_azure.session import Session
