+++ tools/c7n_azure/c7n_azure/resources/vm.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from c7n_azure.actions import AzureBaseAction
+from c7n_azure.actions.base import AzureBaseAction
 from c7n_azure.filters import AzureOffHour, AzureOnHour
 from c7n_azure.provider import resources
 from c7n_azure.resources.arm import ArmResourceManager
