+++ tools/c7n_azure/tests/test_armresource.py
@@ -48,7 +48,7 @@ class ArmResourceTest(BaseTest):
 
     @arm_template('vm.json')
     def test_metric_filter_find(self):
-        with patch('c7n_azure.actions.utcnow') as utc_patch:
+        with patch('c7n_azure.utils.utcnow') as utc_patch:
             utc_patch.return_value = self.get_test_date()
 
             p = self.load_policy({
@@ -71,7 +71,7 @@ class ArmResourceTest(BaseTest):
 
     @arm_template('vm.json')
     def test_metric_filter_find_average(self):
-        with patch('c7n_azure.actions.utcnow') as utc_patch:
+        with patch('c7n_azure.utils.utcnow') as utc_patch:
             utc_patch.return_value = self.get_test_date()
 
             p = self.load_policy({
@@ -94,7 +94,7 @@ class ArmResourceTest(BaseTest):
 
     @arm_template('vm.json')
     def test_metric_filter_not_find(self):
-        with patch('c7n_azure.actions.utcnow') as utc_patch:
+        with patch('c7n_azure.utils.utcnow') as utc_patch:
             utc_patch.return_value = self.get_test_date()
 
             p = self.load_policy({
@@ -117,7 +117,7 @@ class ArmResourceTest(BaseTest):
 
     @arm_template('vm.json')
     def test_metric_filter_not_find_average(self):
-        with patch('c7n_azure.actions.utcnow') as utc_patch:
+        with patch('c7n_azure.utils.utcnow') as utc_patch:
             utc_patch.return_value = self.get_test_date()
 
             p = self.load_policy({
@@ -199,7 +199,7 @@ class ArmResourceTest(BaseTest):
 
     @patch('c7n_azure.query.ResourceQuery.filter',
         return_value=fake_arm_resources)
-    @patch('c7n_azure.actions.DeleteAction.process',
+    @patch('c7n_azure.actions.delete.DeleteAction.process',
         return_value='')
     def test_delete_armresource(self, delete_action_mock, filter_mock):
         p = self.load_policy({
@@ -220,7 +220,7 @@ class ArmResourceTest(BaseTest):
 
     @patch('c7n_azure.query.ResourceQuery.filter',
         return_value=fake_arm_resources)
-    @patch('c7n_azure.actions.DeleteAction.process',
+    @patch('c7n_azure.actions.delete.DeleteAction.process',
         return_value='')
     def test_delete_armresource_specific_name(self, delete_action_mock, filter_mock):
         p = self.load_policy({
